feat: port third part of simple dacs
