--- packages/config/src/projects/other/da-beat/dac/bridges/rhinofi.ts
@@ -1,82 +1,0 @@
-import { ChainId } from '@l2beat/shared-pure'
-import { ProjectDiscovery } from '../../../../../discovery/ProjectDiscovery'
-import { getCommittee } from '../../../../../discovery/starkware'
-import { rhinofi } from '../../../../layer2s/rhinofi'
-import {
-  DaAccessibilityRisk,
-  DaAttestationSecurityRisk,
-  DaExitWindowRisk,
-} from '../../types'
-import { DaBridge } from '../../types/DaBridge'
-import { DacTransactionDataType } from '../../types/DacTransactionDataType'
-import { toUsedInProject } from '../../utils/to-used-in-project'
-
-const discovery = new ProjectDiscovery('deversifi')
-const committee = getCommittee(discovery)
-
-export const rhinofiDac = {
-  id: 'rhinofi-dac-bridge',
-  type: 'DAC',
-  display: {
-    name: 'rhino.fi DAC',
-    slug: 'dac',
-    description: 'Rhino.fi DAC on Ethereum.',
-    links: {
-      websites: ['https://rhino.fi/'],
-      apps: ['https://app.rhino.fi/'],
-      documentation: [
-        'https://docs.rhino.fi/',
-        'https://support.rhino.fi/en/',
-        'https://docs.starkware.co/starkex/index.html',
-      ],
-      explorers: [],
-      repositories: [
-        'https://github.com/starkware-libs/starkex-contracts',
-        'https://github.com/rhinofi',
-      ],
-      socialMedia: [
-        'https://rhino.fi/blog',
-        'https://twitter.com/rhinofi',
-        'https://linkedin.com/company/rhinofi/',
-        'https://youtube.com/c/rhinofi',
-        'https://discord.com/invite/26sXx2KAhy',
-      ],
-    },
-  },
-  contracts: {
-    addresses: [
-      discovery.getContractDetails(
-        'Committee',
-        'Data Availability Committee (DAC) contract verifying data availability claim from DAC Members (via multisig check).',
-      ),
-    ],
-    risks: [],
-  },
-  technology: `## Simple DA Bridge
-    The DA bridge is a smart contract verifying a data availability claim from DAC Members via signature verification.
-    The bridge requires a ${committee.minSigners}/${committee.accounts.length} threshold of signatures to be met before the data commitment is accepted.
-  `,
-  permissions: [
-    {
-      name: 'Committee Members',
-      description: `List of addresses authorized to sign data commitments for the DA bridge.`,
-      accounts: committee.accounts.map((operator) => ({
-        address: operator.address,
-        type: 'EOA',
-      })),
-    },
-  ],
-  chain: ChainId.ETHEREUM,
-  requiredMembers: committee.minSigners,
-  totalMembers: committee.accounts.length,
-  transactionDataType: DacTransactionDataType.StateDiffs,
-  members: {
-    type: 'unknown',
-  },
-  usedIn: toUsedInProject([rhinofi]),
-  risks: {
-    attestations: DaAttestationSecurityRisk.SigVerified(true),
-    accessibility: DaAccessibilityRisk.NotEnshrined,
-    exitWindow: DaExitWindowRisk.Immutable,
-  },
-} satisfies DaBridge
