--- packages/config/src/projects/other/da-beat/dac/bridges/sanko.ts
@@ -1,74 +1,0 @@
-import { ChainId } from '@l2beat/shared-pure'
-import { ProjectDiscovery } from '../../../../../discovery/ProjectDiscovery'
-import { sanko } from '../../../../layer3s/sanko'
-import {
-  DaAccessibilityRisk,
-  DaAttestationSecurityRisk,
-  DaExitWindowRisk,
-} from '../../types'
-import { DaBridge } from '../../types/DaBridge'
-import { DacTransactionDataType } from '../../types/DacTransactionDataType'
-import { toUsedInProject } from '../../utils/to-used-in-project'
-
-const discovery = new ProjectDiscovery('sanko', 'arbitrum')
-
-const DAC = discovery.getContractValue<{
-  membersCount: number
-  requiredSignatures: number
-}>('SequencerInbox', 'dacKeyset')
-const { membersCount, requiredSignatures } = DAC
-
-export const sankoDac = {
-  id: 'sanko-dac-bridge',
-  type: 'DAC',
-  display: {
-    name: 'Sanko DAC',
-    slug: 'dac',
-    description: 'Sanko DAC on Ethereum.',
-    links: {
-      websites: ['https://sanko.xyz/'],
-      apps: ['https://sanko.xyz/bridge', 'https://swap.sanko.xyz'],
-      documentation: ['https://sanko-1.gitbook.io/sanko-mainnet-docs/'],
-      explorers: [
-        'https://sanko-mainnet.calderaexplorer.xyz/',
-        'https://tools.sanko.xyz/',
-        'https://explorer.sanko.xyz/',
-      ],
-      repositories: [],
-      socialMedia: [
-        'https://x.com/SankoGameCorp',
-        'https://discord.gg/Cnz62Vfa2C',
-        'https://t.me/sankogamecorp',
-      ],
-    },
-  },
-  contracts: {
-    addresses: [
-      discovery.getContractDetails(
-        'SequencerInbox',
-        'Main entry point for the Sequencer submitting transaction batches.',
-      ),
-    ],
-    risks: [],
-  },
-  technology: `## Simple DA Bridge
-    The DA bridge is a smart contract verifying a data availability claim from DAC Members via signature verification.
-    The bridge requires a ${requiredSignatures}/${membersCount} threshold of signatures to be met before the data commitment is accepted.
-  `,
-  permissions: [
-    // BLS sigs, not EOAs
-  ],
-  chain: ChainId.ETHEREUM,
-  requiredMembers: requiredSignatures,
-  totalMembers: membersCount,
-  transactionDataType: DacTransactionDataType.TransactionDataCompressed,
-  members: {
-    type: 'unknown',
-  },
-  usedIn: toUsedInProject([sanko]),
-  risks: {
-    attestations: DaAttestationSecurityRisk.NotVerified,
-    accessibility: DaAccessibilityRisk.NotEnshrined,
-    exitWindow: DaExitWindowRisk.Immutable,
-  },
-} satisfies DaBridge
