--- packages/config/src/projects/other/da-beat/dac/bridges/xai.ts
@@ -1,121 +1,0 @@
-import { ChainId } from '@l2beat/shared-pure'
-import { ProjectDiscovery } from '../../../../../discovery/ProjectDiscovery'
-import { xai } from '../../../../layer3s/xai'
-import {
-  DaAccessibilityRisk,
-  DaAttestationSecurityRisk,
-  DaExitWindowRisk,
-} from '../../types'
-import { DaBridge } from '../../types/DaBridge'
-import { DacTransactionDataType } from '../../types/DacTransactionDataType'
-import { toUsedInProject } from '../../utils/to-used-in-project'
-
-const discovery = new ProjectDiscovery('xai', 'arbitrum')
-
-const DAC = discovery.getContractValue<{
-  membersCount: number
-  requiredSignatures: number
-}>('SequencerInbox', 'dacKeyset')
-const { membersCount, requiredSignatures } = DAC
-
-export const xaiDac = {
-  id: 'xai-dac-bridge',
-  type: 'DAC',
-  display: {
-    name: 'Xai DAC',
-    slug: 'dac',
-    description: 'Xai DAC on Ethereum.',
-    links: {
-      websites: ['https://xai.games/'],
-      apps: [],
-      documentation: ['https://xai-foundation.gitbook.io/xai-network/'],
-      explorers: ['https://explorer.xai-chain.net/'],
-      repositories: ['https://github.com/OffchainLabs/nitro'],
-      socialMedia: [
-        'https://twitter.com/xai_games',
-        'https://t.me/XaiSentryNodes',
-        'https://discord.gg/xaigames',
-      ],
-    },
-  },
-  contracts: {
-    addresses: [
-      discovery.getContractDetails(
-        'SequencerInbox',
-        'Main entry point for the Sequencer submitting transaction batches.',
-      ),
-    ],
-    risks: [],
-  },
-  technology: `## Simple DA Bridge
-    The DA bridge is a smart contract verifying a data availability claim from DAC Members via signature verification.
-    The bridge requires a ${requiredSignatures}/${membersCount} threshold of signatures to be met before the data commitment is accepted.
-  `,
-  permissions: [
-    // Members: DAC uses BLS sigs, not EOAs
-    // BLS sigs here: https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das
-    {
-      name: 'Sequencers',
-      accounts: discovery.getPermissionsByRole('Sequencer'),
-      description:
-        'Central actors allowed to submit transaction batches to the Sequencer Inbox.',
-      chain: discovery.chain,
-    },
-    {
-      name: 'RollupOwner',
-      accounts: discovery.getAccessControlRolePermission(
-        'UpgradeExecutor',
-        'EXECUTOR_ROLE',
-      ),
-      description:
-        'Multisig that can upgrade authorized batch posters via the UpgradeExecutor contract.',
-    },
-  ],
-  chain: ChainId.ETHEREUM,
-  requiredMembers: requiredSignatures,
-  totalMembers: membersCount,
-  transactionDataType: DacTransactionDataType.TransactionDataCompressed,
-  members: {
-    type: 'public',
-    list: [
-      {
-        name: 'Xai',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'Ex Populus',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'Rug Radio',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'LayerZero',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'Laguna Games',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'Team Secret',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'Offchain Labs',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-      {
-        name: 'Other Third Party',
-        href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
-      },
-    ],
-  },
-  usedIn: toUsedInProject([xai]),
-  risks: {
-    attestations: DaAttestationSecurityRisk.NotVerified,
-    accessibility: DaAccessibilityRisk.NotEnshrined,
-    exitWindow: DaExitWindowRisk.Immutable,
-  },
-} satisfies DaBridge
