+++ packages/config/src/projects/other/da-beat/dac/rhinofi.ts
@@ -1,34 +1,45 @@
-import { DaEconomicSecurityRisk, DaFraudDetectionRisk } from '../types'
-import { DaLayer } from '../types/DaLayer'
-import { rhinofiDac } from './bridges/rhinofi'
+import { ChainId } from '@l2beat/shared-pure'
+import { ProjectDiscovery } from '../../../../discovery/ProjectDiscovery'
+import { getCommittee } from '../../../../discovery/starkware'
+import { rhinofi } from '../../../layer2s/rhinofi'
+import { DAC } from '../templates/dac-template'
+import { DaAttestationSecurityRisk } from '../types'
+import { DacTransactionDataType } from '../types/DacTransactionDataType'
 
-export const rhinofiLayer: DaLayer = {
-  id: 'rhinofi-dac-layer',
-  type: 'DaLayer',
-  kind: 'DAC',
-  display: {
-    name: 'rhino.fi DAC',
-    slug: 'rhinofi',
-    description:
-      'Set of parties responsible for signing and attesting to the availability of data.',
-    links: {
-      websites: [],
-      documentation: [],
-      repositories: [],
-      apps: [],
-      explorers: [],
-      socialMedia: [],
+const discovery = new ProjectDiscovery('deversifi')
+const committee = getCommittee(discovery)
+
+export const rhinofiDac = DAC({
+  project: rhinofi,
+  risks: {
+    attestations: DaAttestationSecurityRisk.SigVerified(true),
+  },
+  bridge: {
+    contracts: {
+      addresses: [
+        discovery.getContractDetails(
+          'Committee',
+          'Data Availability Committee (DAC) contract verifying data availability claim from DAC Members (via multisig check).',
+        ),
+      ],
+      risks: [],
+    },
+    permissions: [
+      {
+        name: 'Committee Members',
+        description: `List of addresses authorized to sign data commitments for the DA bridge.`,
+        accounts: committee.accounts.map((operator) => ({
+          address: operator.address,
+          type: 'EOA',
+        })),
+      },
+    ],
+    chain: ChainId.ETHEREUM,
+    requiredMembers: committee.minSigners,
+    totalMembers: committee.accounts.length,
+    transactionDataType: DacTransactionDataType.StateDiffs,
+    members: {
+      type: 'unknown',
     },
   },
-  technology: `## Simple Committee
-  The Data Availability Committee (DAC) is a set of trusted parties responsible for storing data off-chain and serving it upon demand. 
-  The security guarantees of DACs depend on the specific setup and can vary significantly based on the criteria for selecting committee members, 
-  their operational transparency, and the mechanisms in place to handle disputes and failures.
-  `,
-  bridges: [rhinofiDac],
-  usedIn: [...rhinofiDac.usedIn],
-  risks: {
-    economicSecurity: DaEconomicSecurityRisk.Unknown,
-    fraudDetection: DaFraudDetectionRisk.NoFraudDetection,
-  },
-}
+})
