+++ packages/config/src/projects/other/da-beat/dac/xai.ts
@@ -1,34 +1,90 @@
-import { DaEconomicSecurityRisk, DaFraudDetectionRisk } from '../types'
-import { DaLayer } from '../types/DaLayer'
-import { xaiDac } from './bridges/xai'
+import { ChainId } from '@l2beat/shared-pure'
+import { ProjectDiscovery } from '../../../../discovery/ProjectDiscovery'
+import { xai } from '../../../layer3s/xai'
+import { DAC } from '../templates/dac-template'
+import { DaEconomicSecurityRisk } from '../types'
+import { DacTransactionDataType } from '../types/DacTransactionDataType'
 
-export const xaiLayer: DaLayer = {
-  id: 'xai-dac-layer',
-  type: 'DaLayer',
-  kind: 'DAC',
-  display: {
-    name: 'XAI DAC',
-    slug: 'xai',
-    description:
-      'Set of parties responsible for signing and attesting to the availability of data.',
-    links: {
-      websites: [],
-      documentation: [],
-      repositories: [],
-      apps: [],
-      explorers: [],
-      socialMedia: [],
+const discovery = new ProjectDiscovery('xai', 'arbitrum')
+
+const dac = discovery.getContractValue<{
+  membersCount: number
+  requiredSignatures: number
+}>('SequencerInbox', 'dacKeyset')
+const { membersCount, requiredSignatures } = dac
+
+export const xaiDac = DAC({
+  project: xai,
+  risks: { economicSecurity: DaEconomicSecurityRisk.OffChainVerifiable },
+  bridge: {
+    contracts: {
+      addresses: [
+        discovery.getContractDetails(
+          'SequencerInbox',
+          'Main entry point for the Sequencer submitting transaction batches.',
+        ),
+      ],
+      risks: [],
+    },
+    permissions: [
+      // Members: DAC uses BLS sigs, not EOAs
+      {
+        name: 'Sequencers',
+        accounts: discovery.getPermissionsByRole('Sequencer'),
+        description:
+          'Central actors allowed to submit transaction batches to the Sequencer Inbox.',
+        chain: discovery.chain,
+      },
+      {
+        name: 'RollupOwner',
+        accounts: discovery.getAccessControlRolePermission(
+          'UpgradeExecutor',
+          'EXECUTOR_ROLE',
+        ),
+        description:
+          'Multisig that can upgrade authorized batch posters via the UpgradeExecutor contract.',
+      },
+    ],
+    chain: ChainId.ETHEREUM,
+    requiredMembers: requiredSignatures,
+    totalMembers: membersCount,
+    transactionDataType: DacTransactionDataType.TransactionDataCompressed,
+    members: {
+      type: 'public',
+      list: [
+        {
+          name: 'Xai',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'Ex Populus',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'Rug Radio',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'LayerZero',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'Laguna Games',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'Team Secret',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'Offchain Labs',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+        {
+          name: 'Other Third Party',
+          href: 'https://xai-foundation.gitbook.io/xai-network/about-xai/xai-protocol/anytrust-revolutionizing-blockchain-infrastructure/data-availability-servers-das',
+        },
+      ],
     },
   },
-  technology: `## Simple Committee
-  The Data Availability Committee (DAC) is a set of trusted parties responsible for storing data off-chain and serving it upon demand. 
-  The security guarantees of DACs depend on the specific setup and can vary significantly based on the criteria for selecting committee members, 
-  their operational transparency, and the mechanisms in place to handle disputes and failures.
-  `,
-  bridges: [xaiDac],
-  usedIn: [...xaiDac.usedIn],
-  risks: {
-    economicSecurity: DaEconomicSecurityRisk.OffChainVerifiable,
-    fraudDetection: DaFraudDetectionRisk.NoFraudDetection,
-  },
-}
+})
