+++ packages/config/src/projects/other/da-beat/index.ts
@@ -11,29 +11,29 @@ import { hychainDac } from './dac/hychain'
 import { immutableXDac } from './dac/immutablex'
 import { l3xDac } from './dac/l3x'
 import { mantleDA } from './dac/mantleDA'
-import { moltenLayer } from './dac/molten'
-import { myriaLayer } from './dac/myria'
-import { playblockLayer } from './dac/playblock'
-import { popapexLayer } from './dac/popapex'
-import { rariLayer } from './dac/rari'
-import { realLayer } from './dac/real'
-import { redsonicLayer } from './dac/redsonic'
-import { reyaLayer } from './dac/reya'
-import { rhinofiLayer } from './dac/rhinofi'
-import { sankoLayer } from './dac/sanko'
-import { sorareLayer } from './dac/sorare'
-import { tanxLayer } from './dac/tanx'
-import { witnessLayer } from './dac/witness'
-import { xaiLayer } from './dac/xai'
-import { xlayerLayer } from './dac/xlayer'
-import { zkfairLayer } from './dac/zkfair'
+import { moltenDac } from './dac/molten'
+import { myriaDac } from './dac/myria'
+import { playblockDac } from './dac/playblock'
+import { popapexDac } from './dac/popapex'
+import { rariDac } from './dac/rari'
+import { realDac } from './dac/real'
+import { redsonicDac } from './dac/redsonic'
+import { reyaDac } from './dac/reya'
+import { rhinofiDac } from './dac/rhinofi'
+import { sankoDac } from './dac/sanko'
+import { sorareDac } from './dac/sorare'
+import { tanxDac } from './dac/tanx'
+import { witnessDac } from './dac/witness'
+import { xaiDac } from './dac/xai'
+import { xlayerDac } from './dac/xlayer'
+import { zkfairDac } from './dac/zkfair'
 import { DaLayer } from './types/DaLayer'
 
 export * from './types'
 
 export const daLayers: DaLayer[] = [
-  rhinofiLayer,
-  zkfairLayer,
+  rhinofiDac,
+  zkfairDac,
   arbitrumNovaDac,
   apexDac,
   astarZkEvmDac,
@@ -41,24 +41,24 @@ export const daLayers: DaLayer[] = [
   celestia,
   ethereum,
   near,
-  xlayerLayer,
-  realLayer,
-  reyaLayer,
-  sorareLayer,
-  tanxLayer,
+  xlayerDac,
+  realDac,
+  reyaDac,
+  sorareDac,
+  tanxDac,
   hychainDac,
-  sankoLayer,
-  myriaLayer,
+  sankoDac,
+  myriaDac,
   edgelessDac,
-  redsonicLayer,
-  witnessLayer,
+  redsonicDac,
+  witnessDac,
   degenDac,
-  rariLayer,
-  xaiLayer,
-  popapexLayer,
-  moltenLayer,
+  rariDac,
+  xaiDac,
+  popapexDac,
+  moltenDac,
   l3xDac,
-  playblockLayer,
+  playblockDac,
   mantleDA,
   eigenDA,
 ]
